Answer clang-tidy: make the exception guarantees honest
The harness promises a game never sees an exception from it. Three findings
said that promise was implemented backwards in places: noexcept means
terminate, not that the caller will catch it.

  * Client::configure and ensure_connected are no longer noexcept. They
    allocate, and a failure there is worth catching in call(), which answers
    the game with a default; marked noexcept it was std::terminate instead.
  * log_write and log_configure keep their noexcept and gain a catch-all, so
    the annotation becomes true: a logger that cannot build a line loses the
    line rather than the process.
  * recv_frame keeps its noexcept and fails soft, since a frame it cannot
    hold ends the connection, which a false return already meant.
  * Both mains now run their body through run() inside a try, so a failed
    allocation reports itself instead of terminating silently.

The remaining findings are resolved by documented acceptance: the two news in
the Client singleton have no recovery path, because the function must return
a reference, and NOLINTNEXTLINE records that rather than hiding it.

.clang-tidy commits the check set - bugprone-* and performance-* - with
swappable-parameters and enum-size off and the reasons written down.

Verified: MSVC x64/Win32 and clang-cl under -WERROR, 5/5 tests, both CLIs
smoke tested, clang-tidy clean under the committed config.

+++ src/session.cpp
@@ -58,10 +58,12 @@ std::string to_text(const Json& value, const std::string& fallback) {
     switch (value.kind()) {
         case Json::Kind::string: return value.as_string();
         case Json::Kind::number:
-            // Only an integral number has a text form Python's str() would agree
-            // with; anything else is more likely a mistake than a name.
-            return value.as_int64() == value.as_double() ? std::to_string(value.as_int64())
-                                                         : fallback;
+            // Only an integral number has a text form worth reading as a name;
+            // anything else is more likely a mistake than a name. The cast says
+            // out loud what the comparison used to do implicitly.
+            return static_cast<double>(value.as_int64()) == value.as_double()
+                       ? std::to_string(value.as_int64())
+                       : fallback;
         default: return fallback;
     }
 }
